meta: begin work porting the system to 32 bit x86
This patch moves a lot build related things into per architecture
directories. It additionally provides support in the build and
toolchain scripts for selecting multiple architectures. This patch
additionally adds minimal scaffolding to be able to complile userland
for i386. Note that lots of functionality is still missing, including
tls and processing dynamic linking relocations.

In addition, the kernel still fails to even compile, let alone function
properly.

+++ libs/libipc/endpoint.cpp
@@ -71,7 +71,7 @@ UniquePtr<Message> Endpoint::wait_for_response_impl(uint32_t type) {
 
 bool Endpoint::send_impl(const Message& message) {
     int fd = m_socket ? m_socket->fd() : -1;
-    return write(fd, &message, message.size) == message.size;
+    return write(fd, &message, message.size) == static_cast<ssize_t>(message.size);
 }
 
 void Endpoint::handle_messages() {
